feat(v8/browser): Rename TryCatch integration to browserApiErrorsIntegration
