+++ dev-packages/browser-integration-tests/suites/public-api/debug/test.ts
@@ -25,7 +25,7 @@ sentryTest('logs debug messages correctly', async ({ getLocalTestUrl, page }) =>
       ? [
           'Sentry Logger [log]: Integration installed: InboundFilters',
           'Sentry Logger [log]: Integration installed: FunctionToString',
-          'Sentry Logger [log]: Integration installed: TryCatch',
+          'Sentry Logger [log]: Integration installed: BrowserApiErrors',
           'Sentry Logger [log]: Integration installed: Breadcrumbs',
           'Sentry Logger [log]: Global Handler attached: onerror',
           'Sentry Logger [log]: Global Handler attached: onunhandledrejection',
