+++ packages/browser/src/integrations/index.ts
@@ -1,5 +1,4 @@
 /* eslint-disable deprecation/deprecation */
-export { TryCatch } from './trycatch';
 export { Breadcrumbs } from './breadcrumbs';
 export { LinkedErrors } from './linkederrors';
 export { HttpContext } from './httpcontext';
